Add "From Zero to RAPIDS" presentation
This version of the "From Zero to RAPIDS: Accelerating Data Science and
Machine Learning Workflows on NVIDIA GPUs" presentation was given on
Friday, January 31st, 2020 as part of the SDSC HPC User Training Series.
RAPIDS is NVIDIA's new Python-based framework for accelerating
end-to-end data science and machine learning pipelines on their GPUs.